Update bulknet.py
Added option to accept command line arguments to output to file as well as stdout. '-h' shows a brief usage, [-f <outputfile>] will duplicate stdout to a file with device name and command that is being sent, if not arguments are sent the script will just print to stdout. Redirection to file only works for show commands.